feat(eval-runtime): 公开 Rubric Judge 宿主接口
