+++ docs/zh/guides/eval-runtime.md
@@ -131,10 +131,85 @@ await reportStore.put(result.report);
 
 随机 `runId` 用于区分执行，并影响 artifact identity，但不属于测量计划。同一份 Definition、Policy、Runtime identity 与显式 seed 会生成相同的 `runContractDigest`。
 
+## 通过内部模型网关接入 Rubric Judge
+
+宿主只提供一次模型调用 port。冻结 prompt、输出解析、1～5 分指标契约、evidence、失败语义和 Evaluator identity 均由 OMK 负责。该 port 不应自行重试；重试、超时、预算、缓存和取消仍由 Core 统一控制。
+
+```ts
+import {
+  createRubricJudgeCriterion,
+  createRubricJudgeEvaluatorDefinition,
+  createRubricJudgeEvaluatorRegistration,
+  createRubricJudgeInstrument,
+  createRubricJudgeMetricDefinition,
+  createRubricJudgeRuntimeConfig,
+  createRuntimeIdentity,
+  type OmkLlmJudgeInvocationPort,
+} from 'oh-my-knowledge/eval-runtime';
+
+const gatewayIdentity = createRuntimeIdentity({
+  implementationId: 'acme.model-gateway/v1',
+  version: '2026.09.03',
+  capabilities: { invocation: 'single-call', cancellation: 'cooperative' },
+  fingerprintFacets: { deploymentRevision: 'sha256:...' },
+});
+const instrument = createRubricJudgeInstrument();
+const judgeRuntime = createRubricJudgeRuntimeConfig({
+  executorId: gatewayIdentity.implementationId,
+  model: 'judge-model',
+  effort: 'low',
+  instrument,
+});
+const criterion = createRubricJudgeCriterion({
+  criterionId: 'correctness',
+  prompt: '法国的首都是哪里？',
+  rubric: '答案必须指出巴黎。',
+});
+
+const invocation: OmkLlmJudgeInvocationPort = {
+  identity: gatewayIdentity,
+  providerCost: { reporting: 'optional' },
+  async invoke(request) {
+    const response = await internalGateway.generate({
+      model: request.model,
+      system: request.system,
+      prompt: request.prompt,
+      signal: request.signal,
+    });
+    return { invocationStatus: 'completed', output: response.text, usage: response.usage };
+  },
+};
+```
+
+把 `criterion` 放入各 sample 的 `evaluationContext` 稳定路径，再把匹配的普通 Core 片段加入 Definition：
+
+```ts
+const evaluator = createRubricJudgeEvaluatorDefinition({
+  evaluatorId: 'correctness-judge',
+  metricId: 'correctness-score',
+  instrument,
+  runtime: judgeRuntime,
+  criterionPointer: '/correctness',
+});
+const metric = createRubricJudgeMetricDefinition('correctness-score');
+
+const runtime = createEvaluationRuntime({
+  executors: [/* 业务 Target registrations */],
+  evaluators: [createRubricJudgeEvaluatorRegistration([{
+    evaluatorId: evaluator.evaluatorId,
+    instrument,
+    runtime: judgeRuntime,
+    invocation,
+  }])],
+});
+```
+
+只有当所有 Target 都产出公共 `SourceNeutralTrace` 契约时，才使用 `tracePolicy: 'source-neutral'`；否则保留默认值 `none`。非 JSON、格式错误、越界分数或缺失理由都会成为结构化 invalid observation，而不是降级成 `0` 分。Provider failure 会保留计量事实并脱敏 provider 私有细节。未注册 Judge 时，不会发现 provider、读取凭证或执行 preflight。
+
 ## 高级宿主
 
 当一个 Definition binding 独占 Core Executor 或 Evaluator port 时，使用 `{ port }` 注册；多个 Target 或 Evaluator 复用同一个 implementation 时，使用 `{ implementationId, createPort }`，让每个 binding 获得隔离的 run 生命周期。若 Definition 声明了 `versionConstraint`，而 registration 没有提供 `satisfiesVersionConstraint`，Runtime 会 fail closed。
 
 自定义进程内 port 时，使用 `createSameProcessExecutorAdapter` 与 `createSameProcessEvaluatorAdapter`。需要分阶段执行、持久化 artifact admission、自定义 Analysis Runtime 或显式跨 run 可比性时，使用 `oh-my-knowledge/eval-core` 的高级 API。参见[嵌入式 Evaluation Core API](/zh/reference/embedded-api)。
 
-完整的可运行 package fixture 位于 [`test/eval-runtime/fixtures/embedded-host.mjs`](https://github.com/lizhiyao/oh-my-knowledge/blob/main/test/eval-runtime/fixtures/embedded-host.mjs)。
+可运行 package fixture 分别覆盖[精确匹配](https://github.com/lizhiyao/oh-my-knowledge/blob/main/test/eval-runtime/fixtures/embedded-host.mjs)和[宿主持有的 Rubric Judge 网关](https://github.com/lizhiyao/oh-my-knowledge/blob/main/test/eval-runtime/fixtures/rubric-judge-host.mjs)。
