+++ src/eval-runtime/index.ts
@@ -56,3 +56,52 @@ export type {
   SameProcessResourceLeaseAccess,
   SameProcessRunScope,
 } from './adapters/same-process.js';
+export type {
+  OmkLlmJudgeEffort,
+  OmkLlmJudgeInvocationPort,
+  OmkLlmJudgeInvocationRequest,
+  OmkLlmJudgeInvocationResult,
+} from './judges/invocation.js';
+export {
+  RUBRIC_JUDGE_BINDINGS,
+  RUBRIC_JUDGE_CONTEXT_SCHEMA,
+  RUBRIC_JUDGE_CONTEXT_SCHEMA_VERSION,
+  RUBRIC_JUDGE_EVALUATOR_IMPLEMENTATION_ID,
+  RUBRIC_JUDGE_EVIDENCE_SCHEMA,
+  RUBRIC_JUDGE_EVIDENCE_SCHEMA_VERSION,
+  RUBRIC_JUDGE_INSTRUMENT_SCHEMA,
+  RUBRIC_JUDGE_INSTRUMENT_SCHEMA_VERSION,
+  createRubricJudgeCriterion,
+  createRubricJudgeEvaluator,
+  createRubricJudgeEvaluatorDefinition,
+  createRubricJudgeEvaluatorIdentity,
+  createRubricJudgeEvaluatorRegistration,
+  createRubricJudgeInstrument,
+  createRubricJudgeMetricDefinition,
+  createRubricJudgeRuntimeConfig,
+  rubricJudgeInstrumentId,
+} from './judges/rubric-judge.js';
+export type {
+  CreateRubricJudgeEvaluatorInput,
+  RubricJudgeConfig,
+  RubricJudgeCriterion,
+  RubricJudgeEvaluatorDefinitionBuilderInput,
+  RubricJudgeEvaluatorBinding,
+  RubricJudgeInstrument,
+  RubricJudgeRuntimeConfig,
+  RubricJudgeTracePolicy,
+} from './judges/rubric-judge.js';
+export {
+  SOURCE_NEUTRAL_TRACE_SCHEMA_DESCRIPTOR,
+  SOURCE_NEUTRAL_TRACE_SCHEMA_VERSION,
+  SOURCE_NEUTRAL_TRACE_WITHOUT_MOCKS_SCHEMA_DESCRIPTOR,
+  SourceNeutralMockStatsSchema,
+  SourceNeutralTraceSchema,
+  SourceNeutralTraceWithoutMocksSchema,
+  attachSourceNeutralMockStats,
+  parseSourceNeutralTrace,
+} from './traces/source-neutral.js';
+export type {
+  SourceNeutralMockStats,
+  SourceNeutralTrace,
+} from './traces/source-neutral.js';
